reorder comms and DGEMM
the goal is to overlap communication and computation.
instead of waitingo on all the GA NB comms then doing all the DGEMMs,
we split the DGEMMs into 3 groups, which have different inputs.
instead of doing the f[1234][nt] pairs back-to-back,
we group the before and after sets, because they have different dependencies.
this allows us to move NB waits down (after DGEMM), which potentially
allows overlap of those comms with the previous DGEMMs.

Signed-off-by: Jeff Hammond <jeff.science@gmail.com>